refactored the store
