+++ src/hocs/with-map/with-map.jsx
@@ -5,6 +5,8 @@ import {connect} from 'react-redux';
 
 import OfferTypes from '../../types/offer.js';
 import {CITIES, MAP_SETTINGS} from '../../helpers/constants.js';
+import {getSelectedCity} from '../../store/reducers/filter/selectors.js';
+import {getActiveOffer} from '../../store/reducers/active/selectors.js';
 
 const withMap = (Component) => {
   class WithMap extends React.PureComponent {
@@ -112,8 +114,8 @@ const withMap = (Component) => {
   }
 
   const mapStateToProps = (state) => ({
-    selectedCity: state.selectedCity,
-    activeOffer: state.activeOffer
+    selectedCity: getSelectedCity(state),
+    activeOffer: getActiveOffer(state)
   });
 
   WithMap.propTypes = {
