+++ src/index.js
@@ -7,7 +7,8 @@ import {compose} from 'recompose';
 
 import App from './components/app/app.jsx';
 import reviews from './mocks/reviews.js';
-import {reducer, Operation} from './store/reducer.js';
+import reducer from './store/reducer.js';
+import Operation from './store/operations/data/data.js';
 import createAPI from './api/api.js';
 
 const api = createAPI((...args) => store.dispatch(...args));
