+++ src/store/actions/data/data.js
@@ -1,0 +1,14 @@
+export const Actions = {
+  getOffers: `GET_OFFERS`,
+};
+
+const ActionCreator = {
+
+  getOffers: (offers) => ({
+    type: Actions.getOffers,
+    payload: offers
+  }),
+
+};
+
+export default ActionCreator;
