+++ src/store/actions/details/details.js
@@ -1,0 +1,14 @@
+export const Actions = {
+  setDetailsOffer: `SET_DETAILS_OFFER`
+};
+
+const ActionCreator = {
+
+  setDetailsOffer: (offer) => ({
+    type: Actions.setDetailsOffer,
+    payload: offer
+  })
+
+};
+
+export default ActionCreator;
