+++ src/store/name-space.js
@@ -1,0 +1,8 @@
+const NameSpace = {
+  DATA: `DATA`,
+  FILTER: `FILTER`,
+  ACTIVE: `ACTIVE`,
+  DETAILS: `DETAILS`,
+};
+
+export default NameSpace;
