+++ src/store/operations/data/data.js
@@ -1,0 +1,14 @@
+import ActionCreator from '../../actions/data/data.js';
+import {formatOffers} from '../../../helpers/helpers.js';
+
+const Operation = {
+  loadOffers: () => (dispatch, getState, api) => {
+    return api.get(`/hotels`)
+      .then(
+          (response) => {
+            dispatch(ActionCreator.getOffers(formatOffers(response.data)));
+          });
+  }
+};
+
+export default Operation;
