+++ src/store/reducers/active/active.e2e.test.js
@@ -1,0 +1,31 @@
+import {Actions} from '../../actions/active/active.js';
+import reducer from './active.js';
+
+describe(`Active reducer work correctly`, () => {
+
+  it(`Should change activeOffer with given value`, () => {
+    const state1 = {
+      activeOffer: null
+    };
+    const action = {
+      type: Actions.setActiveOffer,
+      payload: {id: 1, title: `title1`, city: `Amsterdam`}
+    };
+    expect(reducer(state1, action)).toMatchObject({
+      activeOffer: {id: 1, title: `title1`, city: `Amsterdam`}
+    });
+  });
+
+  it(`Should remove activeOffer `, () => {
+    const state2 = {
+      activeOffer: {id: 1, title: `title1`, city: `Amsterdam`}
+    };
+    const action = {
+      type: Actions.removeActiveOffer
+    };
+    expect(reducer(state2, action)).toMatchObject({
+      activeOffer: null
+    });
+  });
+
+});
