+++ src/store/reducers/data/selectors.js
@@ -1,0 +1,7 @@
+import NameSpace from '../../name-space.js';
+
+const NAME_SPACE = NameSpace.DATA;
+
+export const getOffers = (state) => {
+  return state[NAME_SPACE].offers;
+};
